backend/pyproject.toml
[tool.poetry]
name = "sashreek"
version = "0.1.0"
description = ""
authors = ["Leo Trubach <leotrubach@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2.7"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
psycopg = "^3.1.12"
python-dotenv = "^1.0.0"
dj-database-url = "^2.1.0"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.3.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
ruff = "^0.1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
